Merge test progress to master branch

+++ controllers/GroupController.py
@@ -47,7 +47,7 @@ class GroupController:
     def print_choices(self):
         print(
             """
----------🅼 🅴 🅽 🆄------------
+-----------🅼 🅴 🅽 🆄------------
 0. Go Back
 1. Create group
 2. Update group
@@ -65,35 +65,49 @@ class GroupController:
 """
         )
 
+    # Function for creating a group
     def create_group(self):
         name = input("Enter group name: ")
         group = CreateGroupDto(name=name)
-        group_id = self.group_service.create_group(group)
-        group = self.group_service.view_group(group_id)
+        group_id = self.group_service.create_group(group) #Create group from GroupService
+        group = self.group_service.view_group(group_id) #Get the new group's details
 
+        #Display the new group's details
         print("\nSuccessfully created your group!")
         print("Here are the details: ")
         print("\nGroup ID: " + str(group.id))
         print("Name: " + group.name)
         print("Date Created: " + str(group.dateCreated))
 
+    #Function for updating group name
     def update_group(self):
         _id = int(input("Enter group id: "))
-        name = input("Enter group name: ")
-        group = CreateGroupDto(name=name)
-        self.group_service.update_group(_id, group)
-        print("Successfully renamed your group to " + name)
+        group = self.group_service.view_group(_id)
 
+        if group != None:
+            name = input("Enter group name: ")
+            group = CreateGroupDto(name=name)
+            self.group_service.update_group(_id, group)
+            print("Successfully renamed your group to " + name) 
+
+    #Function for deleting a group by id
     def delete_group(self):
         _id = int(input("Enter group id: "))
-        self.group_service.delete_group(_id)
-        print("Successfully deleted group")
+        group = self.group_service.view_group(_id)
 
+        if group != None:
+            self.group_service.delete_group(_id)
+            print("Successfully deleted group")
+
+    #Function for viewing all groups
     def view_all_groups(self):
         groups = self.group_service.view_all_groups()
+
+        #Check groups exists
         if len(groups) == 0:
             print("There are no existing groups")
-        else:
+        else: 
+            #Print details of each group
             print("\nAll groups: \n")
             for group in groups:
                 print("Group ID: " + str(group["id"]))
@@ -101,65 +115,89 @@ class GroupController:
                 print("Date created: " + str(group["date"]))
                 print("------------------------------------")
 
+    #Function for adding a friend to a group
     def add_person_to_group(self):
+        #Get group id
         group_id = int(input("Enter group id: "))
-        person_id = int(input("Enter person id: "))
+        #Get all friends
         members = self.group_service.view_all_persons_in_group(group_id)
-        friend = self.friend_service.get_friend_by_id(person_id)
         group = self.group_service.view_group(group_id)
 
-        isFriendAlreadyInGroup = False
-        for member in members:
-            if member.id != person_id:
-                continue
-            else:
-                isFriendAlreadyInGroup = True
-                break
+        if group != None: #Check if group exist
+            person_id = int(input("Enter person id: "))
+            friend = self.friend_service.get_friend_by_id(person_id)
+            #Check if friend is already in group
+            isFriendAlreadyInGroup = False
+            for member in members:
+                if member.id != person_id:
+                    continue
+                else:
+                    isFriendAlreadyInGroup = True
+                    break
 
-        if isFriendAlreadyInGroup:
-             print(friend.name + " is already in " + group.name)
-        else:
-            self.group_service.add_person_to_group(person_id, group_id)
-            print("Successfully added " + friend.name + " to " + group.name)
+            
+            if isFriendAlreadyInGroup:
+                print(friend.name + " is already in " + group.name)
+            else:
+                if friend == None:
+                    print("Friend does not exist!")
+                else:
+                    self.group_service.add_person_to_group(person_id, group_id)
+                    print("Successfully added " + friend.name + " to " + group.name)
            
 
     def remove_person_from_group(self):
         group_id = int(input("Enter group id: "))
-        person_id = int(input("Enter person id: "))
         members = self.group_service.view_all_persons_in_group(group_id)
-        friend = self.friend_service.get_friend_by_id(person_id)
         group = self.group_service.view_group(group_id)
 
-        isFriendInGroup = False
-        for member in members:
-            if member.id != person_id:
-                continue
-            else:
-                isFriendInGroup = True
-                break
+        if group != None: #Check if group exists
+            person_id = int(input("Enter person id: "))
+            friend = self.friend_service.get_friend_by_id(person_id)
 
-        if isFriendInGroup:
-            self.group_service.remove_person_from_group(person_id, group_id)
-            print("Successfully removed " + friend.name + " to " + group.name)
-        else:
-            print(friend.name + " is not in " + group.name)
+            #Check if friend is in the group
+            isFriendInGroup = False
+            for member in members:
+                if member.id != person_id:
+                    continue
+                else:
+                    isFriendInGroup = True
+                    break
+
+            if isFriendInGroup:
+                self.group_service.remove_person_from_group(person_id, group_id)
+                print("Successfully removed " + friend.name + " to " + group.name)
+            else:
+                if friend == None: #Check if friend exists
+                    print("Friend does not exist!")
+                else:
+                    print(friend.name + " is not in " + group.name)
 
     def view_group(self):
         _id = int(input("Enter group id: "))
         group = self.group_service.view_group(_id)
-        print("\nGroup ID: " + str(group.id))
-        print("Name: " + group.name)
-        print("Date Created: " + str(group.dateCreated))
-        print("------------------------------------")
+
+        if group != None: #Print details if group exists
+            print("\nGroup ID: " + str(group.id))
+            print("Name: " + group.name)
+            print("Date Created: " + str(group.dateCreated))
+            print("------------------------------------")
 
     def view_all_members(self):
         _id = int(input("Enter group id: "))
-        print("\nMembers in Group with ID: " + str(_id) + "\n")
-        members = self.group_service.view_all_persons_in_group(_id)
-        for member in members:
-            print("Person ID: " + str(member.id))
-            print("Name: " + member.name)
-            print("------------------------------------")
+        group = self.group_service.view_group(_id)
+
+        if group != None: #Check if group exists
+            print("\nMembers in Group with ID: " + str(_id) + "\n")
+            members = self.group_service.view_all_persons_in_group(_id)
+
+            if len(members) == 0:  #Check if there are members in the group
+                print("There are no members in this group")
+            else: 
+                for member in members:
+                    print("Person ID: " + str(member.id))
+                    print("Name: " + member.name)
+                    print("------------------------------------")
 
     def remove_members(self):
         print("\nChoose Group ID to delete all members:\n")
@@ -174,9 +212,14 @@ class GroupController:
         for group in choices:
             counter += 1
             if str(group["id"]) == str(_id):
-                self.group_service.remove_all_persons_from_group(int(_id))
-                print("Successfully removed all members from " + group["name"])
-                break
+                members = self.group_service.view_all_persons_in_group(int(_id))
+
+                if len(members) == 0: #Check if there are members in the group
+                    print("There are no members in this group")
+                else: 
+                    self.group_service.remove_all_persons_from_group(int(_id))
+                    print("Successfully removed all members from " + group["name"])
+                    break
             elif counter == len(choices):
                 print("Group ID not found.")
 
@@ -191,10 +234,19 @@ class GroupController:
 
     def view_persons_groups(self):
         person_id = int(input("Enter person id: "))
-        groups = self.group_service.view_all_groups_of_person(person_id)
-        print("\nGroups joined by a person with id: " + str(person_id) + "\n")
-        for group in groups:
-            print("Group ID: " + str(group.id))
-            print("Name: " + group.name)
-            print("Date Created: " + str(group.dateCreated))
-            print("------------------------------------")
+        friend = self.friend_service.get_friend_by_id(person_id)
+
+        if friend != None: #check if friend exists
+            groups = self.group_service.view_all_groups_of_person(person_id)
+
+            if len(groups) == 0:
+                print("Person does not belong to any group")
+            else:
+                print("\nGroups joined by a person with id: " + str(person_id) + "\n")
+                for group in groups:
+                    print("Group ID: " + str(group.id))
+                    print("Name: " + group.name)
+                    print("Date Created: " + str(group.dateCreated))
+                    print("------------------------------------")
+        else: 
+            print("Friend does not exist!")
